% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circular.R
\name{hdg_circular}
\alias{hdg_circular}
\title{Use functions from the circular package}
\usage{
hdg_circular(hdg)
}
\arguments{
\item{hdg}{A heading in degrees, where 0 is north,
90 is east, 180 is south, and 270 is west. Values
outside the range [0-360) are coerced to this range
using \code{\link[=hdg_norm]{hdg_norm()}}.}
}
\value{
A \code{\link[circular:circular]{circular::circular()}} object.
}
\description{
Use functions from the circular package
}
\examples{
if (requireNamespace("circular", quietly = TRUE)) {
  hdg_circular(1:10)
}

}
