% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heading.R
\name{hdg_decl}
\alias{hdg_decl}
\alias{hdg_true_from_magnetic}
\alias{hdg_magnetic_from_true}
\title{Calculate declination, true heading, and magnetic heading}
\usage{
hdg_decl(
  lon,
  lat,
  year,
  height = 0,
  height_ref = c("geoid", "ellipsoid"),
  model = c("IGRF13", "WMM2020", "EMM2017")
)

hdg_true_from_magnetic(
  hdg,
  lon,
  lat,
  year,
  height = 0,
  height_ref = c("geoid", "ellipsoid"),
  model = c("IGRF13", "WMM2020", "EMM2017")
)

hdg_magnetic_from_true(
  hdg,
  lon,
  lat,
  year,
  height = 0,
  height_ref = c("geoid", "ellipsoid"),
  model = c("IGRF13", "WMM2020", "EMM2017")
)
}
\arguments{
\item{lon, lat}{Vectors of coordinates.}

\item{year}{A decimal year (e.g., 2020.2120). The World Magnetic Model
is valid between 2020.0 and 2025.0; the Enhanced Magnetic Model
is valid between 2000.0 and 2022.0.}

\item{height}{A height above the earth's surface (as approximated
by the EGM9615 geoid) in kilometers. If \code{height_ref} is
"ellipsoid", this is interpreted as the height above the
WGS84 ellipsoid.}

\item{height_ref}{One of geoid or ellipsoid.}

\item{model}{One of IGRF13, WMM2020, or EMM2017.}

\item{hdg}{A heading in degrees, where 0 is north,
90 is east, 180 is south, and 270 is west. Values
outside the range [0-360) are coerced to this range
using \code{\link[=hdg_norm]{hdg_norm()}}.}
}
\value{
A declination, true heading, or magnetic heading.
}
\description{
For more detailed output, see \code{\link[=wmm2020_extract]{wmm2020_extract()}}.
}
\examples{
hdg_decl(-64, 45, year = 2021)
hdg_true_from_magnetic(13.40, -64, 45, year = 2021)
hdg_magnetic_from_true(356.51, -64, 45, year = 2021)

}
