% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heading.R
\name{hdg_norm}
\alias{hdg_norm}
\alias{uv}
\alias{uv_norm}
\alias{uv_from_hdg}
\alias{hdg_from_uv}
\alias{rad_from_hdg}
\alias{hdg_from_rad}
\title{Create and export headings}
\usage{
hdg_norm(hdg)

uv(u, v)

uv_norm(uv)

uv_from_hdg(hdg)

hdg_from_uv(uv)

rad_from_hdg(hdg)

hdg_from_rad(rad)
}
\arguments{
\item{hdg}{A heading in degrees, where 0 is north,
90 is east, 180 is south, and 270 is west. Values
outside the range [0-360) are coerced to this range
using \code{\link[=hdg_norm]{hdg_norm()}}.}

\item{uv, u, v}{A data.frame with columns \code{u} (magnitude east)
and \code{v} (magnitude north).}

\item{rad}{An angle in radians such that a \code{hdg} of 90 is
zero radians and a \code{hdg} of 0 is \code{pi / 2} radians.}
}
\description{
Create and export headings
}
\examples{
hdg_norm(-10:10)

}
