% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kamloops2016}
\alias{kamloops2016}
\title{Kamloops Hourly Climate Data}
\format{
An object of class \code{data.frame} with 8784 rows and 37 columns.
}
\usage{
kamloops2016
}
\description{
Obtained from \href{https://climate.weather.gc.ca/historical_data/search_historic_data_e.html}{Environment Canada}
via the \href{https://docs.ropensci.org/weathercan/}{weathercan}
package.
}
\examples{
head(kamloops2016[c("time", "wind_dir", "wind_spd")])

}
\keyword{datasets}
