% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heading.R
\name{hdg_mean}
\alias{hdg_mean}
\alias{hdg_diff}
\alias{hdg_sd}
\title{Heading arithmetic}
\usage{
hdg_mean(hdg, na.rm = FALSE)

hdg_diff(hdg, hdg_ref)

hdg_sd(hdg, na.rm = FALSE)
}
\arguments{
\item{hdg}{A heading in degrees, where 0 is north,
90 is east, 180 is south, and 270 is west. Values
outside the range [0-360) are coerced to this range
using \code{\link[=hdg_norm]{hdg_norm()}}.}

\item{na.rm}{Use \code{TRUE} to remove missing values}

\item{hdg_ref}{A reference heading against which
to compare \code{hdg}.}
}
\description{
Heading arithmetic
}
\examples{
hdg_mean(-10:10)
hdg_mean(c(350, 10))

hdg_diff(350:370, 0)

}
