% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headings-package.R
\docType{package}
\name{headings-package}
\alias{headings}
\alias{headings-package}
\title{headings: Tools for Working With Compass Headings}
\description{
Provides functions for manipulating compass headings.
}
\author{
\strong{Maintainer}: Dewey Dunnington \email{dewey.dunnington@dfo-mpo.gc.ca} (\href{https://orcid.org/0000-0002-9415-4582}{ORCID})

}
\keyword{internal}
