% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magnetic-models.R
\docType{data}
\name{wmm2025_extract}
\alias{wmm2025_extract}
\alias{wmm2020_extract}
\alias{igrf14_extract}
\alias{igrf13_extract}
\alias{emm2017_extract}
\alias{mm_ellipsoidal_height}
\alias{mm_decimal_year}
\alias{mm_version}
\alias{mm_egm9615_geoid_int}
\title{IGRF and NCEI World/Enhanced Magnetic Model Interface}
\format{
An object of class \code{integer} of length 1038961.
}
\usage{
wmm2025_extract(
  lon,
  lat,
  year = mm_decimal_year(Sys.Date()),
  height = mm_ellipsoidal_height(lon, lat, 0)
)

wmm2020_extract(
  lon,
  lat,
  year = mm_decimal_year(Sys.Date()),
  height = mm_ellipsoidal_height(lon, lat, 0)
)

igrf14_extract(
  lon,
  lat,
  year = mm_decimal_year(Sys.Date()),
  height = mm_ellipsoidal_height(lon, lat, 0)
)

igrf13_extract(
  lon,
  lat,
  year = mm_decimal_year(Sys.Date()),
  height = mm_ellipsoidal_height(lon, lat, 0)
)

emm2017_extract(
  lon,
  lat,
  year = mm_decimal_year(Sys.Date()),
  height = mm_ellipsoidal_height(lon, lat, 0)
)

mm_ellipsoidal_height(lon, lat, height)

mm_decimal_year(date)

mm_version()

mm_egm9615_geoid_int
}
\arguments{
\item{lon, lat}{Vectors of coordinates.}

\item{year}{A decimal year (e.g., 2020.2120). The World Magnetic Model
is valid between 2020.0 and 2025.0; the Enhanced Magnetic Model
is valid between 2000.0 and 2022.0.}

\item{height}{Height above the WGS84 ellipsoid in kilometers. Use
\code{\link[=mm_ellipsoidal_height]{mm_ellipsoidal_height()}} to convert geoid elevations (EGM9615)
to ellipsoidal values.}

\item{date}{An R date object or value that can be coerced to one.}
}
\value{
A data frame with columns - \code{decl}: Magnetic variation in degrees -
\code{incl}: Inclination of the magnetic field in degrees - \code{decl_err},
\code{incl_err}: Error associated with declination and inclination,
respectively.
}
\description{
IGRF and NCEI World/Enhanced Magnetic Model Interface
}
\examples{
wmm2025_extract(-64, 45, mm_decimal_year("2026-01-01"))
wmm2020_extract(-64, 45, mm_decimal_year("2021-01-01"))
emm2017_extract(-64, 45, mm_decimal_year("2021-01-01"))
igrf14_extract(-64, 45, mm_decimal_year("2026-01-01"))
igrf13_extract(-64, 45, mm_decimal_year("2021-01-01"))

}
\references{
NCEI World Magnetic Model \url{https://www.ngdc.noaa.gov/geomag/WMM/DoDWMM.shtml}.

NCEI Enhanced Magnetic Model \url{https://www.ngdc.noaa.gov/geomag/EMM/index.html}.

IGRF-13 \url{https://www.ngdc.noaa.gov/IAGA/vmod/home.html}.

NCEI Geomagnetic Modeling Team and British Geological Survey.
2019. World Magnetic Model 2020. NOAA National Centers for Environmental
Information. \doi{10.25921/11v3-da71}, 2020, accessed 2021-02-23.

Chulliat, A., W. Brown, P. Alken, C. Beggan, M. Nair, G. Cox, A. Woods, S.
Macmillan, B. Meyer and M. Paniccia, The US/UK World Magnetic Model for
2020-2025: Technical Report, National Centers for Environmental Information,
NOAA, \doi{10.25923/ytk1-yx35}, 2020.
}
\keyword{datasets}
