% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmm.R
\name{wmm_extract}
\alias{wmm_extract}
\alias{wmm_read_coefficients}
\alias{wmm_decimal_year}
\alias{wmm_version}
\title{NCEI World Magnetic Model 2020}
\usage{
wmm_extract(
  lon,
  lat,
  year = wmm_decimal_year(Sys.Date()),
  height = 0,
  coef = NULL
)

wmm_read_coefficients(file)

wmm_decimal_year(date)

wmm_version()
}
\arguments{
\item{lon, lat}{Coordinates in WGS84 datum}

\item{year}{A decimal year (e.g., 2020.2120). The World Magnetic Model
is valid between 2020 and 2025}

\item{height}{Height above the WGS84 ellipsoid in kilometers.}

\item{coef}{Model coefficients as generated by \code{\link[=wmm_read_coefficients]{wmm_read_coefficients()}}, or
\code{NULL} for the default WMM 2020 values.}

\item{file}{Path to a ".COF" file containing coefficients in the form
required by the model.}

\item{date}{An R date object or value that can be coerced to one.}
}
\value{
A data frame with columns - \code{decl}: Magnetic variation in degrees -
\code{incl}: Inclination of the magnetic field in degrees - \code{decl_err},
\code{incl_err}: Error associated with declination and inclination,
respectively.
}
\description{
NCEI World Magnetic Model 2020
}
\examples{
wmm_version()
wmm_extract(-64, 45, wmm_decimal_year("2021-01-01"))

}
\references{
NCEI World Magnetic Model \url{https://www.ngdc.noaa.gov/geomag/WMM/DoDWMM.shtml}.

NCEI Geomagnetic Modeling Team and British Geological Survey.
2019. World Magnetic Model 2020. NOAA National Centers for Environmental
Information. \doi{10.25921/11v3-da71}, 2020, accessed 2021-02-23.

Chulliat, A., W. Brown, P. Alken, C. Beggan, M. Nair, G. Cox, A. Woods, S.
Macmillan, B. Meyer and M. Paniccia, The US/UK World Magnetic Model for
2020-2025: Technical Report, National Centers for Environmental Information,
NOAA, \doi{10.25923/ytk1-yx35}, 2020.
}
